requirements.txt
pandas==2.2.3
numpy==1.26.4
faiss-cpu==1.9.0
streamlit==1.40.1
nltk==3.9.1
langchain==0.3.7
langchain-openai==0.2.8
openai==1.54.4
